% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{get_correlations}
\alias{get_correlations}
\title{Retrieve within- and/or between-cluster correlations for a wbCorr object (tables).
Correlations matrices can be obtained via the \code{summary()} or \code{summary.wbCorr()} function.}
\usage{
get_correlations(object, which = c("within", "between"))
}
\arguments{
\item{object}{A wbCorr object, created by the \code{wbCorr()} function.}

\item{which}{A character vector indicating which correlation table to return.
Options are \code{'within'} or \code{'w'}, and \code{'between'} or \code{'b'}.}
}
\value{
A list containing the selected tables of within- and/or between-cluster correlations.
}
\description{
Retrieve within- and/or between-cluster correlations for a wbCorr object (tables).
Correlations matrices can be obtained via the \code{summary()} or \code{summary.wbCorr()} function.
}
\examples{
# Example using the iris dataset
wbCorrObject <- wbCorr(iris, iris$Species)
correlations <- get_correlations(wbCorrObject, which = c('within', 'between', 'merge'))
correlations$within
correlations$between

}
\seealso{
\code{\link[=summary.wbCorr]{summary}}, \code{\link[=wbCorr]{wbCorr}}
}
