% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within-between-correlations.R
\name{summary,wbCorr-method}
\alias{summary,wbCorr-method}
\alias{summary.wbCorr}
\title{Summarize within- and/or between-cluster correlations for a wbCorr object.}
\usage{
\S4method{summary}{wbCorr}(object, which = c("within", "between", "merge"), ...)
}
\arguments{
\item{object}{A wbCorr object, created by the \code{wbCorr()} function.}

\item{which}{A string or a character vector indicating which summaries to return.
Options are \code{'within'} or \code{'w'}, \code{'between'} or \code{'b'}, and various merge options
like \code{'merge'}, \code{'m'}, \code{'merge_wb'}, \code{'wb'}, \code{'merge_bw'}, \code{'bw'}.
Default is \code{c('within', 'between', 'merge')}.}

\item{...}{Additional arguments passed to the base summary method}
}
\value{
A list containing the selected summaries of within- and/or between-cluster correlations.
}
\description{
Summarize within- and/or between-cluster correlations for a wbCorr object.
}
\examples{
# Example using the iris dataset
cors <- wbCorr(iris, iris$Species)
summaries <- summary(cors, which = c('within', 'between', 'merge'))
summaries$within
summaries$between
summaries$merged_wb

}
\seealso{
\code{\link[=get_correlations]{get_correlations}}, \code{\link[=wbCorr]{wbCorr}}
}
