% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_update.R
\name{update_my_package}
\alias{update_my_package}
\title{Update the R package to the latest version on GitHub}
\usage{
update_my_package(repo, current_version)
}
\arguments{
\item{repo}{A string with the GitHub repository in the format "username/repository".}

\item{current_version}{A string with the current package version to compare against.}
}
\value{
A boolean value indicating if the package was updated (TRUE) or not (FALSE).
}
\description{
This function checks if there is a newer version of the package available on GitHub,
and if so, updates the package to the latest version.
}
\examples{
update_my_package("yourusername/yourpackage", "1.0.0")
}
