% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_post_02_accessor_functions.R
\name{get_ICC}
\alias{get_ICC}
\alias{get_ICCs}
\title{Return all ICCs for the original variables.}
\usage{
get_ICC(object)

get_ICCs(object)
}
\arguments{
\item{object}{A wbCorr object, created by the wbCorr() function.}
}
\value{
A dataframe with ICCs for all variables. ICC is obtained by fitting
mixed effects models and extracting the variance components. Then, the formula
between- variance / total- variance is applied.
}
\description{
You can use get_ICC() or get_ICCs() interchangeably.
}
\examples{
# importing our simulated example dataset with pre-specified within- and between- correlations
data("simdat_intensive_longitudinal")

# create object:
correlations <- wbCorr(data = simdat_intensive_longitudinal,
                      cluster = 'participantID')

# returns the ICCs:
ICCs <- get_ICC(correlations)
print(ICCs)

}
\seealso{
\code{\link[=wbCorr]{wbCorr}}
}
