% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R,
%   R/within-between-correlations.R
\name{get_matrix}
\alias{get_matrix}
\alias{get_matrices}
\alias{summary,wbCorr-method}
\alias{summary.wbCorr}
\title{Return matrices for within- and/or between-cluster correlations.}
\usage{
get_matrix(object, which = c("within", "between", "merge"), ...)

get_matrices(object, which = c("within", "between", "merge"), ...)

\S4method{summary}{wbCorr}(object, which = c("within", "between", "merge"), ...)
}
\arguments{
\item{object}{A wbCorr object, created by the wbCorr() function.}

\item{which}{A string or a character vector indicating which summaries to return.
Options are 'within' or 'w', 'between' or 'b', and various merge options
like 'merge', 'm', 'merge_wb', 'wb', 'merge_bw', 'bw'.
Default is c('within', 'between', 'merge').}

\item{...}{Additional arguments passed to the base summary method}
}
\value{
A list containing the selected matrices of within- and/or between-cluster correlations.
}
\description{
You can use summary(), get_matrices(), or get_matrix() interchangeably.
}
\examples{
# Example using the iris dataset
cors <- wbCorr(iris, iris$Species)
matrices <- get_matrices(cors, which = c('within', 'between', 'merge'))
matrices$within
matrices$between
matrices$merged_wb

}
\seealso{
\code{\link[=get_table]{get_tables}}, \code{\link[=wbCorr]{wbCorr}}
}
