% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{get_table}
\alias{get_table}
\alias{get_tables}
\title{Retrieve full tables for both within- and/or between-cluster correlations for a wbCorr object.}
\usage{
get_table(object, which = c("within", "between"))

get_tables(object, which = c("within", "between"))
}
\arguments{
\item{object}{A wbCorr object, created by the wbCorr() function.}

\item{which}{A character vector indicating which correlation table to return.
Options are 'within' or 'w', and 'between' or 'b'.}
}
\value{
A list containing the selected tables of within- and/or between-cluster correlations.
}
\description{
This function has an alias get_tables() which can be used interchangeably.
For correlations matrices, see the summary() function.
}
\examples{
# importing our simulated example dataset with pre-specified within- and between- correlations
data("simdat_intensive_longitudinal")

# create object:
correlations <- wbCorr(data = simdat_intensive_longitudinal,
                      cluster = 'participantID')

# returns a list with full detailed tables of the correlations:
tables <- get_table(correlations) # the get_tables() function is equivalent
print(tables)

# Access specific tables by:
# Option 1:
tables$between
# Option 2:
within_table <- get_tables(correlations, which = 'w') # or use 'within' or 'between'
print(within_table) # within_table could be saved to an excel or csv file (e.g., write.csv)

}
\seealso{
\code{\link[=summary.wbCorr]{summary}}, \code{\link[=wbCorr]{wbCorr}}
}
