% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{get_table}
\alias{get_table}
\alias{get_tables}
\title{Retrieve full tables for both within- and/or between-cluster correlations for a wbCorr object.}
\usage{
get_table(object, which = c("within", "between"))

get_tables(object, which = c("within", "between"))
}
\arguments{
\item{object}{A wbCorr object, created by the wbCorr() function.}

\item{which}{A character vector indicating which correlation table to return.
Options are 'within' or 'w', and 'between' or 'b'.}
}
\value{
A list containing the selected tables of within- and/or between-cluster correlations.
}
\description{
This function has an alias get_table() which can be used interchangeably.
For correlations matrices, use the summary() function.
}
\examples{
# Example using the iris dataset
wbCorrObject <- wbCorr(iris, iris$Species)
tables <- get_tables(wbCorrObject, which = c('within', 'between'))
tables$within
tables$between
}
\seealso{
\code{\link[=summary.wbCorr]{summary}}, \code{\link[=wbCorr]{wbCorr}}
}
