% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbCorr.R
\name{print,wbCorr-method}
\alias{print,wbCorr-method}
\alias{print.wbCorr}
\title{Print Method for the wbCorr Class}
\usage{
\S4method{print}{wbCorr}(x, ...)
}
\arguments{
\item{x}{A \code{wbCorr} object.}

\item{...}{Additional arguments, currently unused.}
}
\description{
Prints a summary of the \code{wbCorr} object.
}
\examples{
# Example using the iris dataset
cors <- wbCorr(iris, iris$Species, weighted_between_statistics = TRUE)
print(cors)
}
\seealso{
\code{\link[=wbCorr]{wbCorr}}
}
