% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_export_to_excel.R
\name{to_excel}
\alias{to_excel}
\title{Saves the passed summary or table to excel}
\usage{
to_excel(SummaryObject, path = file.path(getwd(), "wbCorr.xlsx"))
}
\arguments{
\item{SummaryObject}{A summary or matrix object, such as those returned by \code{get_matrix()} or \code{get_table()}.}

\item{path}{Specify the filename and a path. If no path is provided, the file will be saved to the
current working directory.}
}
\value{
Writes an Excel file (.xlsx) to disk.
}
\description{
Use \code{to_excel(get_matrix(wbCorrObject))} or \code{to_excel(get_table(wbCorrObject))} to save
the provided table/matrix to an excel file.
}
\examples{
# Importing our simulated example dataset with pre-specified within- and between- correlations
data("simdat_intensive_longitudinal")

# Create object:
correlations <- wbCorr(data = simdat_intensive_longitudinal,
                      cluster = 'participantID')

# Returns a correlation matrix with stars for p-values:
matrices <- get_matrix(correlations) # summary(correlations) works too.

to_excel(matrices)

}
\seealso{
\code{\link[=get_table]{get_tables}}, \code{\link[=wbCorr]{wbCorr}}, \code{\link[=get_matrix]{get_matrix}}
}
