% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{update_wbCorr}
\alias{update_wbCorr}
\title{Check for package updates and update if available}
\usage{
update_wbCorr()
}
\value{
Returns TRUE if an update is available, FALSE otherwise.
}
\description{
This function checks for updates of the package from its GitHub repository
and prompts the user to update the package if a new version is available.
}
