% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbCorr.R
\docType{class}
\name{wbCorr-class}
\alias{wbCorr-class}
\title{wbCorr Class}
\description{
A class representing within- and between-cluster correlations.
}
\details{
The \code{wbCorr} class is used to store within- and between-cluster correlations
and provides methods for printing and summarizing the correlations.
}
\seealso{
\code{\link[=wbCorr]{wbCorr}}
}
