% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbCorr.R
\name{wbCorr}
\alias{wbCorr}
\title{wbCorr}
\usage{
wbCorr(
  data,
  cluster,
  alpha_level = 0.95,
  method = "pearson",
  weighted_between_statistics = FALSE
)
}
\arguments{
\item{data}{A dataframe containing numeric variables for which correlations will be calculated.}

\item{cluster}{A vector representing the clustering variable or a string with the name of the column in \code{data} that contains the clustering variable.}

\item{alpha_level}{A numeric value between 0 and 1 representing the desired level of confidence for confidence intervals (default: 0.95).}

\item{method}{A string indicating the correlation method to be used. Supported methods are 'pearson', 'kendall', and 'spearman' (default: 'pearson').}

\item{weighted_between_statistics}{A logical value. If False, Variables are centered between-persons by
simply taking the mean for each person and weighting them all the same, even if some
contributed less measurement points. If TRUE, correlations are weighted. These methods will be equivalen in datasets
without missing data and an equal amount of measurements per person. TRUE only supports continuous variables (default: FALSE)}
}
\value{
A wbCorr object that contains within- and between-cluster correlations, p-values, and confidence intervals.
Use the \code{get_table()} function on the wbCorr object to retrieve full tables.
Use the \code{summary()} or \code{get_matrix()} function on the wbCorr object to retrieve correlation matrices.
}
\description{
The \code{wbCorr} function creates a wbCorr object containing within- and between-cluster correlations,
p-values, and confidence intervals for a given dataset and clustering variable.
}
\details{
Calculates within- and between-cluster correlations for a given dataset and clustering variable.

Inspired by the psych::statsBy function, \code{wbCorr} allows you to easily calculate and extract within-
and between-cluster correlations for further analysis.
}
\examples{
# Example using the iris dataset
cors <- wbCorr(iris, iris$Species, weighted_between_statistics = TRUE)
cors
get_tables(cors, which = c('within', 'between'))
summary(cors, which = c('within', 'between', 'merge'))
}
\seealso{
\code{\link[=get_table]{get_table}}, \code{\link[=summary.wbCorr]{summary}}, \code{\link[=get_matrix]{get_matrix}}
}
