% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbCorr.R
\name{wbCorr}
\alias{wbCorr}
\alias{wbcorr}
\title{wbCorr}
\usage{
wbCorr(
  data,
  cluster,
  confidence_level = 0.95,
  method = "pearson",
  bootstrap = FALSE,
  nboot = 1000,
  weighted_between_statistics = FALSE
)

wbcorr(
  data,
  cluster,
  confidence_level = 0.95,
  method = "pearson",
  bootstrap = FALSE,
  nboot = 1000,
  weighted_between_statistics = FALSE
)
}
\arguments{
\item{data}{A dataframe containing numeric variables for which correlations will be calculated.}

\item{cluster}{A vector representing the clustering variable or a string with the name of the column in data that contains the clustering variable.}

\item{confidence_level}{A numeric value between 0 and 1 representing the desired level of confidence for confidence intervals (default: 0.95).}

\item{method}{A string indicating the correlation method to be used.
Supported methods are 'pearson', 'spearman', 'spearman-jackknife', and 'auto'.
(default: 'pearson'). 'pearson': Pearson correlation method uses t-statistics
to determine confidence intervals and p-values.'spearman': Spearman correlation
method uses the Fisher z-transformation for confidence intervals and p-values (unless bootstrap is enabled),
assuming normally distributed data. 'spearman-jackknife': Spearman-Jackknife
correlation method employs the Euclidean jackknife technique to compute
confidence intervals, providing more robust confidence intervals in the presence of
non-normal data or outliers. Note that p-values are not available
when this method is selected. 'auto' uses pearson for numeric variables and
spearman for correlations involving at least one factors. Still check your
assumptions.}

\item{bootstrap}{Performs a bias-corrected and accelerated (BCa) bootstrap to compute both confidence
intervals, as well as p-values (default: FALSE). Recommended, but slow.}

\item{nboot}{Specifies the amount of bootstrap samples. We recommend a minimum of 1000 (default: 1000).}

\item{weighted_between_statistics}{A logical value. If FALSE, variables are centered between persons by
simply taking the mean for each person and weighting them all the same, even if some
contributed fewer measurement points. If TRUE, correlations are weighted. These methods will be equivalent in datasets
without missing data and an equal number of measurements per person. TRUE only supports continuous variables (default: FALSE).}
}
\value{
A wbCorr object that contains within- and between-cluster statistics.
Use the get_table() function on the wbCorr object to retrieve a list of the full correlation tables.
Use the summary() or get_matrix() function on the wbCorr object to retrieve various correlation matrices.
Use  get_ICC() in order to get all intra class correlations.
}
\description{
The wbCorr function creates a wbCorr object containing within- and between-cluster correlations,
p-values, and confidence intervals for a given dataset and clustering variable.
}
\details{
Calculates within- and between-cluster correlations for a given dataset and clustering variable.

Inspired by the psych::statsBy function, wbCorr allows you to easily calculate and extract within-
and between-cluster correlations for further analysis.
}
\examples{
# importing our simulated example dataset with pre-specified within- and between- correlations
data("simdat_intensive_longitudinal")
# use ?simdat_intensive_longitudinal # documentation of the dataset and the "true" correlations
head(simdat_intensive_longitudinal)

# returns an object:
correlations <- wbCorr(data = simdat_intensive_longitudinal,
cluster = 'participantID')

# returns a list with full detailed tables of the correlations:
tables <- get_table(correlations) # the get_tables() function is equivalent
print(tables)

# returns a correlation matrix with stars for p-values:
matrices <- summary(correlations) # the get_matrix() and get_matrices() functions are equivalent
print(matrices)

# Access specific tables or matrices by:
# Option 1:
matrices$within
tables$between
# Option 2:
get_tables(correlations, which = 'within')
summary(correlations, which = c('w', 'wb', 'bw')) # abbreviations equivalent to full words

}
\seealso{
\code{\link[=get_table]{get_table}}, \code{\link[=summary.wbCorr]{summary}}, \code{\link[=get_ICC]{get_ICC}}
}
